feat(validate) added date, date-time, time format attribute validation
This adds the ability to validate the format attribute within a schema
via a regular expression for `date`, `date-time` and `time`. `date-time`
is the only added attribute in this commit that is listed in Draft 4;
however, the draft indicates custom formats may exist. This added
validation follows the RFC3339 specification section 5.6 and section 5.7
for dates and time.